+++ web/src/components/TopNav.tsx
@@ -72,120 +72,118 @@ function TopNav() {
 
   return (
     <>
-      <Navbar expand="lg" bg="dark" sticky="top">
-        <Container>
-          <OverlayTrigger placement="bottom" overlay={popover}>
-            <Navbar.Brand>
-              <img
-                alt=""
-                src={logo}
-                width="32"
-                height="32"
-                className="d-inline-block align-top"
-              />
-            </Navbar.Brand>
-          </OverlayTrigger>
-          <Navbar.Toggle aria-controls="top-nav" />
-          <Navbar.Collapse id="top-nav" role="">
-            <Nav className="me-auto">
-              <LinkContainer to="commands">
-                <Nav.Link>
-                  <i className="fa-solid fa-terminal pe-1"></i>Commands
-                </Nav.Link>
-              </LinkContainer>
-              <LinkContainer to="streams">
-                <Nav.Link>
-                  <i className="fa-brands fa-twitch pe-1"></i>ALttP Streams
-                </Nav.Link>
-              </LinkContainer>
-              <LinkContainer to="twitch">
-                <Nav.Link>
-                  <i className="fa-solid fa-robot pe-1"></i>Twitch Bot
-                </Nav.Link>
-              </LinkContainer>
-              <LinkContainer to="discord">
-                <Nav.Link>
-                  <i className="fa-brands fa-discord pe-1"></i>Discord Bot
-                </Nav.Link>
-              </LinkContainer>
-              {user && user.permissions.includes("admin") && (
-                <LinkContainer to="admin">
-                  <Nav.Link>
-                    <i className="fa-solid fa-user-tie pe-1"></i>Admin
-                  </Nav.Link>
-                </LinkContainer>
-              )}
-            </Nav>
-            <Nav>
-              <Nav.Link
-                href="https://github.com/greenham/helpasaur-king"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <i className="fa-brands fa-github pe-1"></i>GitHub
+      <Navbar expand="lg" bg="dark" sticky="top" className="px-3">
+        <OverlayTrigger placement="bottom" overlay={popover}>
+          <Navbar.Brand>
+            <img
+              alt=""
+              src={logo}
+              width="32"
+              height="32"
+              className="d-inline-block align-top"
+            />
+          </Navbar.Brand>
+        </OverlayTrigger>
+        <Navbar.Toggle aria-controls="top-nav" />
+        <Navbar.Collapse id="top-nav" role="">
+          <Nav className="me-auto">
+            <LinkContainer to="commands">
+              <Nav.Link>
+                <i className="fa-solid fa-terminal pe-1"></i>Commands
               </Nav.Link>
+            </LinkContainer>
+            <LinkContainer to="streams">
+              <Nav.Link>
+                <i className="fa-brands fa-twitch pe-1"></i>ALttP Streams
+              </Nav.Link>
+            </LinkContainer>
+            <LinkContainer to="twitch">
+              <Nav.Link>
+                <i className="fa-solid fa-robot pe-1"></i>Twitch Bot
+              </Nav.Link>
+            </LinkContainer>
+            <LinkContainer to="discord">
+              <Nav.Link>
+                <i className="fa-brands fa-discord pe-1"></i>Discord Bot
+              </Nav.Link>
+            </LinkContainer>
+            {user && user.permissions.includes("admin") && (
+              <LinkContainer to="admin">
+                <Nav.Link>
+                  <i className="fa-solid fa-user-tie pe-1"></i>Admin
+                </Nav.Link>
+              </LinkContainer>
+            )}
+          </Nav>
+          <Nav>
+            <Nav.Link
+              href="https://github.com/greenham/helpasaur-king"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <i className="fa-brands fa-github pe-1"></i>GitHub
+            </Nav.Link>
+            <NavDropdown
+              title={
+                <>
+                  <i className="fa-solid fa-link pe-1"></i>Resources
+                </>
+              }
+              id="resources-dropdown"
+            >
+              {RESOURCES.map((resource, index) =>
+                resource.divider ? (
+                  <NavDropdown.Divider key={index} />
+                ) : (
+                  <NavDropdown.Item
+                    key={index}
+                    href={resource.href}
+                    target={resource.target}
+                    rel={resource.rel}
+                  >
+                    <i className={`${resource.icon} pe-1`}></i>
+                    {resource.text}
+                  </NavDropdown.Item>
+                )
+              )}
+            </NavDropdown>
+          </Nav>
+          <Nav className="justify-content-end">
+            {user ? (
               <NavDropdown
                 title={
-                  <>
-                    <i className="fa-solid fa-link pe-1"></i>Resources
-                  </>
+                  <Image
+                    src={user.twitchUserData.profile_image_url}
+                    roundedCircle
+                    className="bg-dark ml-3"
+                    alt={"Logged in as " + user.twitchUserData.display_name}
+                    width={32}
+                    height={32}
+                  />
                 }
-                id="resources-dropdown"
+                id="user-dropdown"
               >
-                {RESOURCES.map((resource, index) =>
-                  resource.divider ? (
-                    <NavDropdown.Divider key={index} />
-                  ) : (
-                    <NavDropdown.Item
-                      key={index}
-                      href={resource.href}
-                      target={resource.target}
-                      rel={resource.rel}
-                    >
-                      <i className={`${resource.icon} pe-1`}></i>
-                      {resource.text}
-                    </NavDropdown.Item>
-                  )
-                )}
-              </NavDropdown>
-            </Nav>
-            <Nav className="justify-content-end">
-              {user ? (
-                <NavDropdown
-                  title={
-                    <Image
-                      src={user.twitchUserData.profile_image_url}
-                      roundedCircle
-                      className="bg-dark ml-3"
-                      alt={"Logged in as " + user.twitchUserData.display_name}
-                      width={32}
-                      height={32}
-                    />
+                <NavDropdown.Item
+                  href={
+                    process.env.API_HOST +
+                    "/auth/logout?redirect=" +
+                    encodeURIComponent(location.pathname)
                   }
-                  id="user-dropdown"
+                  rel="noopener noreferrer"
                 >
-                  <NavDropdown.Item
-                    href={
-                      process.env.API_HOST +
-                      "/auth/logout?redirect=" +
-                      encodeURIComponent(location.pathname)
-                    }
-                    rel="noopener noreferrer"
-                  >
-                    <i className="fa-solid fa-arrow-right-from-bracket"></i>
-                    Log Out
-                  </NavDropdown.Item>
-                </NavDropdown>
-              ) : (
-                <Nav.Link href={getTwitchLoginUrl()} rel="noopener noreferrer">
-                  <Button variant="primary">
-                    <i className="fa-solid fa-key pe-1"></i>Log in with Twitch
-                  </Button>
-                </Nav.Link>
-              )}
-            </Nav>
-          </Navbar.Collapse>
-        </Container>
+                  <i className="fa-solid fa-arrow-right-from-bracket"></i>
+                  Log Out
+                </NavDropdown.Item>
+              </NavDropdown>
+            ) : (
+              <Nav.Link href={getTwitchLoginUrl()} rel="noopener noreferrer">
+                <Button variant="primary">
+                  <i className="fa-solid fa-key pe-1"></i>Log in with Twitch
+                </Button>
+              </Nav.Link>
+            )}
+          </Nav>
+        </Navbar.Collapse>
       </Navbar>
     </>
   );